src/main/java/com/codegym/wdbsspringboot/service/taskservice/TaskService.java
package com.codegym.wdbsspringboot.service.taskservice;

import com.codegym.wdbsspringboot.model.Task;
import com.codegym.wdbsspringboot.repository.ITaskRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class TaskService implements ITaskService{

    @Autowired
    private ITaskRepository taskRepository;

    @Override
    public Iterable<Task> findAll() {
        return taskRepository.findAll();
    }

    @Override
    public Optional<Task> findById(Long id) {
        return taskRepository.findById(id);
    }

    @Override
    public Task save(Task task) {
        return taskRepository.save(task);
    }

    @Override
    public void remove(Long id) {
        taskRepository.deleteById(id);
    }

    public TaskService(ITaskRepository repository) {
        this.taskRepository = repository;
    }
}
